requirements.txt
flask
pdfminer.six
scikit-learn
nltk